[K6.0] Try to fix the ajax controller to make raw /json request working
