+++ src/site/src/Controller/Topic/Json/TopicJsonDisplay.php
@@ -1,0 +1,117 @@
+<?php
+/**
+ * Kunena Component
+ *
+ * @package         Kunena.Site
+ * @subpackage      Controller.Category
+ *
+ * @copyright       Copyright (C) 2008 - 2022 Kunena Team. All rights reserved.
+ * @license         https://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @link            https://www.kunena.org
+ **/
+
+namespace Kunena\Forum\Site\Controller\Topic\Json;
+
+\defined('_JEXEC') or die();
+
+use Exception;
+use Joomla\CMS\Factory;
+use Joomla\CMS\Filesystem\File;
+use Joomla\CMS\Language\Text;
+use Joomla\CMS\MVC\Controller\BaseController;
+use Joomla\CMS\Uri\Uri;
+use Kunena\Forum\Libraries\Access\KunenaAccess;
+use Kunena\Forum\Libraries\Controller\KunenaControllerDisplay;
+use Kunena\Forum\Libraries\Forum\Category\KunenaCategory;
+use Kunena\Forum\Libraries\Forum\Category\KunenaCategoryHelper;
+use Kunena\Forum\Libraries\Forum\Message\KunenaMessageHelper;
+use Kunena\Forum\Libraries\Forum\Topic\KunenaTopic;
+use Kunena\Forum\Libraries\Forum\Topic\KunenaTopicHelper;
+use Kunena\Forum\Libraries\Pagination\KunenaPagination;
+use Kunena\Forum\Libraries\Route\KunenaRoute;
+use Kunena\Forum\Libraries\User\KunenaUser;
+use Kunena\Forum\Libraries\User\KunenaUserHelper;
+use Kunena\Forum\Site\Model\CategoryModel;
+
+/**
+ * Class ComponentKunenaControllerApplicationMiscDisplay
+ *
+ * @since   Kunena 4.0
+ */
+class TopicJsonDisplay extends KunenaControllerDisplay
+{
+	/**
+	 * @var     string
+	 * @since   Kunena 6.0
+	 */
+	public $headerText;
+
+	/**
+	 * @var     KunenaCategory
+	 * @since   Kunena 6.0
+	 */
+	public $category;
+
+	/**
+	 * @var     integer
+	 * @since   Kunena 6.0
+	 */
+	public $total;
+
+	/**
+	 * @var     KunenaTopic
+	 * @since   Kunena 6.0
+	 */
+	public $topics;
+
+	/**
+	 * @var     KunenaPagination
+	 * @since   Kunena 6.0
+	 */
+	public $pagination;
+
+	/**
+	 * @var     KunenaUser
+	 * @since   Kunena 6.0
+	 */
+	public $me;
+
+	/**
+	 * @var     string
+	 * @since   Kunena 6.0
+	 */
+	protected $name = 'Topic/Json';
+
+	/**
+	 * Prepare category display.
+	 *
+	 * @return  void
+	 *
+	 * @throws  null
+	 * @since   Kunena 6.0
+	 */
+	protected function before()
+	{
+	    $catid    = $this->app->input->getInt('catid', 0);
+	    
+	    $category = KunenaCategoryHelper::get($catid);
+	    
+	    $response = $category->topictemplate;
+
+	    return $response;
+	}
+
+	/**
+	 * Prepare document.
+	 *
+	 * @return  void
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 * @since   Kunena 6.0
+	 */
+	protected function prepareDocument()
+	{
+		
+	}
+}
